Update server.php
Please add this code to fix header corruption issues for windows users.  This fixes invalid or unexpected token error messages in the browser if this code is not added to the file.  This will make it easier for newcomers to use Laravel since this error is hard to troubleshoot for beginners.